# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Sentinel API
  version: 0.2.0
  description: Fundação e autenticação por senha da Fase 1.
servers:
  - url: http://localhost:8080
paths:
  /health/live:
    get:
      operationId: healthLive
      responses:
        "200":
          description: Processo ativo
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Health"
  /health/ready:
    get:
      operationId: healthReady
      responses:
        "200":
          description: PostgreSQL disponível e serviço pronto para receber tráfego
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Health"
        "503":
          description: Dependência necessária indisponível
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
  /v1/auth/register:
    post:
      operationId: register
      summary: Cria uma conta; não cria sessão.
      parameters:
        - $ref: "#/components/parameters/Origin"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CredentialsRequest" }
      responses:
        "201":
          description: Conta criada
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "400": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
        "429": { $ref: "#/components/responses/Problem" }
  /v1/auth/login:
    post:
      operationId: login
      summary: Cria uma sessão por senha e rotaciona uma sessão apresentada.
      parameters:
        - $ref: "#/components/parameters/Origin"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CredentialsRequest" }
      responses:
        "200":
          description: Login concluído
          headers:
            Set-Cookie:
              schema: { type: string }
              description: Cookie __Host-session Secure, HttpOnly, SameSite=Lax e Path=/.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SessionResponse" }
        "401": { $ref: "#/components/responses/Problem" }
        "429": { $ref: "#/components/responses/Problem" }
  /v1/auth/me:
    get:
      operationId: currentSession
      security: [{ sessionCookie: [] }]
      responses:
        "200":
          description: Sessão ativa
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SessionResponse" }
        "401": { $ref: "#/components/responses/Problem" }
  /v1/auth/csrf:
    get:
      operationId: issueCsrf
      security: [{ sessionCookie: [] }]
      responses:
        "200":
          description: CSRF emitido ou renovado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CsrfResponse" }
        "401": { $ref: "#/components/responses/Problem" }
  /v1/auth/logout:
    post:
      operationId: logout
      security: [{ sessionCookie: [] }]
      parameters:
        - $ref: "#/components/parameters/Origin"
        - name: X-CSRF-Token
          in: header
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Sessão revogada e cookie removido
          headers:
            Set-Cookie:
              schema: { type: string }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-session
  parameters:
    Origin:
      name: Origin
      in: header
      required: true
      schema: { type: string, format: uri }
  responses:
    Problem:
      description: Erro RFC 9457
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/ProblemDetails" }
  headers:
    RequestId:
      description: UUID de correlação validado ou gerado pelo servidor.
      schema:
        type: string
        format: uuid
  schemas:
    Health:
      type: object
      required: [status, service]
      properties:
        status: { type: string, const: ok }
        service: { type: string, const: sentinel-api }
    ProblemDetails:
      type: object
      required: [type, title, status, code, correlation_id]
      properties:
        type: { type: string, format: uri-reference }
        title: { type: string }
        status: { type: integer, minimum: 400, maximum: 599 }
        code: { type: string }
        correlation_id: { type: string, format: uuid }
    CredentialsRequest:
      type: object
      additionalProperties: false
      required: [email, password]
      properties:
        email: { type: string, format: email, maxLength: 254 }
        password: { type: string, minLength: 15, maxLength: 1024, writeOnly: true }
    User:
      type: object
      required: [id, email]
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
    SessionDetails:
      type: object
      required: [id, idle_expires_at, absolute_expires_at]
      properties:
        id: { type: string, format: uuid }
        idle_expires_at: { type: string, format: date-time }
        absolute_expires_at: { type: string, format: date-time }
    SessionResponse:
      type: object
      required: [user, session]
      properties:
        user: { $ref: "#/components/schemas/User" }
        session: { $ref: "#/components/schemas/SessionDetails" }
    CsrfResponse:
      type: object
      required: [csrf_token, expires_at]
      properties:
        csrf_token: { type: string, writeOnly: true }
        expires_at: { type: string, format: date-time }